docker_autotag/requirements_for_ocr.txt
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
einops==0.8.1
entmax==1.3
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pix2tex==0.1.4
pydantic==2.11.3
pydantic_core==2.33.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
timm==0.5.4
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
x-transformers==0.15.0
